requirements.txt
asttokens==3.0.0
decorator==5.2.0
executing==2.2.0
iniconfig==2.0.0
ipython==8.32.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
numpy==2.2.3
opt_einsum==3.4.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
scipy==1.15.2
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
wheel==0.45.1
